+++ Sources/Sentry/include/SentryTracer.h
@@ -3,7 +3,7 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-@class SentryHub, SentryTransactionContext, SentryTraceHeader, SentryTraceState,
+@class SentryHub, SentryTransactionContext, SentryTraceHeader, SentryTraceContext,
     SentryDispatchQueueWrapper;
 
 static NSTimeInterval const SentryTracerDefaultTimeout = 3.0;
@@ -47,7 +47,7 @@ static NSTimeInterval const SentryTracerDefaultTimeout = 3.0;
 /**
  * Retrieves a trace state from this tracer.
  */
-@property (nonatomic, readonly) SentryTraceState *traceState;
+@property (nonatomic, readonly) SentryTraceContext *traceContext;
 
 /*
  The root span of this tracer.